requirements.txt
discord.py==2.2.0
python-dotenv==1.0.0
requests==2.28.2
emoji==2.2.0